Refactor: implemented a reusable vbox validator composable
